+++ crates/cli-sub-agent/src/cli.rs
@@ -208,6 +208,14 @@ pub enum Commands {
         #[arg(long)]
         no_stream_stdout: bool,
 
+        /// Disable the fatal-error-marker silent-hang scan for this session. Use
+        /// ONLY when developing CSA's own error/quota/failover detection code,
+        /// whose source and test fixtures legitimately contain provider error
+        /// markers (#1745). The idle-timeout and wall-clock timeout still apply.
+        /// Default: scan enabled.
+        #[arg(long)]
+        no_error_marker_scan: bool,
+
         /// Path to agent-spec file (.spec or .toml) for contract-based verification
         #[arg(long, value_name = "PATH", value_parser = parse_spec_path_arg)]
         spec: Option<String>,
