+++ crates/cli-sub-agent/src/goal_loop.rs
@@ -104,6 +104,9 @@ pub(crate) struct GoalRunRequest {
     pub(crate) tier: Option<String>,
     pub(crate) force_ignore_tier_setting: bool,
     pub(crate) no_fs_sandbox: bool,
+    /// CLI `--no-error-marker-scan`: disables the #1652 fatal-error-marker
+    /// silent-hang scan for this run, overriding config (#1745).
+    pub(crate) no_error_marker_scan: bool,
     pub(crate) extra_writable: Vec<PathBuf>,
     pub(crate) extra_readable: Vec<PathBuf>,
 }
@@ -169,6 +172,7 @@ pub(crate) async fn handle_run_or_goal(request: GoalRunRequest) -> Result<i32> {
         request.tier,
         request.force_ignore_tier_setting,
         request.no_fs_sandbox,
+        request.no_error_marker_scan,
         request.extra_writable,
         request.extra_readable,
     )
@@ -326,6 +330,7 @@ async fn run_goal_iteration(
         request.tier.clone(),
         request.force_ignore_tier_setting,
         request.no_fs_sandbox,
+        request.no_error_marker_scan,
         request.extra_writable.clone(),
         request.extra_readable.clone(),
     )
