+++ crates/cli-sub-agent/src/pipeline_tests_locking.rs
@@ -55,6 +55,7 @@ async fn execute_with_session_and_meta_does_not_persist_runtime_binary_when_lock
         false,
         &[],
         &[],
+        false, // cli_no_error_marker_scan (#1745)
     )
     .await;
     let err = match execution {
