+++ crates/cli-sub-agent/src/pipeline_tests_pre_session_hook.rs
@@ -104,6 +104,7 @@ printf 'ok\n'
         false, // readonly_project_root
         &[],
         &[],
+        false, // cli_no_error_marker_scan (#1745)
     )
     .await
     .unwrap();
