+++ crates/cli-sub-agent/src/pipeline_tests_tail.rs
@@ -687,7 +687,8 @@ async fn execute_with_session_and_meta_rejects_illegal_result_path_in_real_flow(
         false, // no_fs_sandbox
         false, // readonly_project_root
         &[],   // extra_writable
-        &[],   // extra_readable
+        &[],   // extra_readable,
+        false, // cli_no_error_marker_scan (#1745)
     )
     .await
     .unwrap();
@@ -761,7 +762,8 @@ async fn execute_with_session_and_meta_explicit_only_ignores_inherited_parent_se
         false, // no_fs_sandbox
         false, // readonly_project_root
         &[],   // extra_writable
-        &[],   // extra_readable
+        &[],   // extra_readable,
+        false, // cli_no_error_marker_scan (#1745)
     )
     .await
     .unwrap();
