+++ crates/cli-sub-agent/src/review_cmd_exact_tests.rs
@@ -464,6 +464,7 @@ async fn execute_review_marks_unavailable_when_all_tier_models_fail() {
         false,
         &[],
         &[],
+        true, // no_error_marker_scan: default scan ON (#1745)
     )
     .await
     .expect("all-failed fallback should still return an outcome");
@@ -541,6 +542,7 @@ async fn execute_review_falls_back_to_next_tier_model_and_persists_routing_metad
         false,
         &[],
         &[],
+        true, // no_error_marker_scan: default scan ON (#1745)
     )
     .await
     .expect("tier fallback should succeed");
