+++ crates/cli-sub-agent/src/review_cmd_execute_tests.rs
@@ -81,7 +81,8 @@ printf 'tool mutation\\n' >> \"{}\"\n",
         false, // no_fs_sandbox
         false, // readonly_project_root
         &[],   // extra_writable
-        &[],   // extra_readable
+        &[],   // extra_readable,
+        true,  // no_error_marker_scan: default scan ON (#1745)
     )
     .await
     .expect("review should succeed after reclassifying edit restriction");
@@ -185,6 +186,7 @@ printf '%s\\n' \
         false,
         &[],
         &[],
+        true, // no_error_marker_scan: default scan ON (#1745)
     )
     .await
     .expect("codex review should honor project target");
@@ -278,7 +280,8 @@ printf '%s\\n' \
         false, // no_fs_sandbox
         false, // readonly_project_root
         &[],   // extra_writable
-        &[],   // extra_readable
+        &[],   // extra_readable,
+        true,  // no_error_marker_scan: default scan ON (#1745)
     )
     .await
     .expect("explicit review model spec should bypass tier enforcement");
@@ -465,6 +468,7 @@ fi\n",
         false,
         &[],
         &[],
+        true, // no_error_marker_scan: default scan ON (#1745)
     )
     .await
     .expect("gemini auth retry should succeed");
@@ -532,6 +536,7 @@ async fn execute_review_classifies_gemini_oauth_prompt_without_api_key() {
         false,
         &[],
         &[],
+        true, // no_error_marker_scan: default scan ON (#1745)
     )
     .await
     .expect("classified auth failure should return a result");
@@ -626,6 +631,7 @@ printf 'Opening authentication page\\nDo you want to continue? [Y/n]\\n'\n",
         true,
         &[],
         &[],
+        true, // no_error_marker_scan: default scan ON (#1745)
     )
     .await
     .expect("classified auth failure should still return a result");
