+++ crates/cli-sub-agent/src/run_cmd_attempt.rs
@@ -407,6 +407,7 @@ pub(crate) async fn execute_run_loop(request: RunLoopRequest<'_>) -> Result<RunL
                     request.no_fs_sandbox,
                     &request.extra_writable,
                     &request.extra_readable,
+                    request.no_error_marker_scan,
                 )
                 .await
             }
@@ -449,6 +450,7 @@ pub(crate) async fn execute_run_loop(request: RunLoopRequest<'_>) -> Result<RunL
                 request.no_fs_sandbox,
                 &request.extra_writable,
                 &request.extra_readable,
+                request.no_error_marker_scan,
             )
             .await
         }?;
