+++ crates/cli-sub-agent/src/run_cmd_tests_core.rs
@@ -343,6 +343,32 @@ fn test_cli_no_memory_flag_parses() {
 }
 
 #[test]
+fn test_cli_no_error_marker_scan_flag_parses() {
+    // #1745: opt-out flag must parse on `csa run` and default to false.
+    let cli = try_parse_cli(&["csa", "run", "--no-error-marker-scan", "prompt"]).unwrap();
+    match cli.command {
+        crate::cli::Commands::Run {
+            no_error_marker_scan,
+            ..
+        } => {
+            assert!(no_error_marker_scan);
+        }
+        _ => panic!("expected Run command"),
+    }
+
+    let cli_default = try_parse_cli(&["csa", "run", "prompt"]).unwrap();
+    match cli_default.command {
+        crate::cli::Commands::Run {
+            no_error_marker_scan,
+            ..
+        } => {
+            assert!(!no_error_marker_scan);
+        }
+        _ => panic!("expected Run command"),
+    }
+}
+
+#[test]
 fn test_cli_prompt_flag_parses_and_conflicts_with_positional_prompt() {
     let cli = try_parse_cli(&["csa", "run", "--prompt", "hello"]).unwrap();
     match cli.command {
