+++ crates/cli-sub-agent/src/run_cmd_tier_tests.rs
@@ -153,6 +153,7 @@ async fn handle_run_persists_result_for_direct_tool_tier_rejection() {
         None,
         false,
         false,
+        false, // no_error_marker_scan (#1745)
         Vec::new(),
         Vec::new(),
     )
