+++ crates/csa-executor/src/transport_tests_gemini_fallback.rs
@@ -118,6 +118,7 @@ async fn test_execute_falls_back_to_api_key_after_all_retries_exhausted() {
         output_spool: None,
         output_spool_max_bytes: csa_process::DEFAULT_SPOOL_MAX_BYTES,
         output_spool_keep_rotated: csa_process::DEFAULT_SPOOL_KEEP_ROTATED,
+        error_marker_scan_enabled: true,
         setting_sources: None,
         sandbox: None,
         thinking_budget: None,
@@ -297,6 +298,7 @@ async fn test_execute_best_effort_sandbox_fallback_preserves_attempt_model_overr
         output_spool: None,
         output_spool_max_bytes: csa_process::DEFAULT_SPOOL_MAX_BYTES,
         output_spool_keep_rotated: csa_process::DEFAULT_SPOOL_KEEP_ROTATED,
+        error_marker_scan_enabled: true,
         setting_sources: None,
         sandbox: Some(&sandbox),
         thinking_budget: None,
