+++ crates/csa-executor/src/transport_tests_gemini_sandbox_lockstep.rs
@@ -76,6 +76,7 @@ async fn test_execute_fails_fast_when_shared_npm_cache_bind_cannot_be_added() {
         output_spool: None,
         output_spool_max_bytes: csa_process::DEFAULT_SPOOL_MAX_BYTES,
         output_spool_keep_rotated: csa_process::DEFAULT_SPOOL_KEEP_ROTATED,
+        error_marker_scan_enabled: true,
         setting_sources: None,
         sandbox: Some(&sandbox),
         thinking_budget: None,
@@ -180,6 +181,7 @@ async fn test_legacy_execute_fails_fast_when_shared_npm_cache_bind_cannot_be_add
         output_spool: None,
         output_spool_max_bytes: csa_process::DEFAULT_SPOOL_MAX_BYTES,
         output_spool_keep_rotated: csa_process::DEFAULT_SPOOL_KEEP_ROTATED,
+        error_marker_scan_enabled: true,
         setting_sources: None,
         sandbox: Some(&sandbox),
         thinking_budget: None,
@@ -284,6 +286,7 @@ async fn test_execute_fails_fast_when_shared_npm_cache_path_violates_writable_al
         output_spool: None,
         output_spool_max_bytes: csa_process::DEFAULT_SPOOL_MAX_BYTES,
         output_spool_keep_rotated: csa_process::DEFAULT_SPOOL_KEEP_ROTATED,
+        error_marker_scan_enabled: true,
         setting_sources: None,
         sandbox: Some(&sandbox),
         thinking_budget: None,
@@ -395,6 +398,7 @@ async fn test_legacy_execute_fails_fast_when_shared_npm_cache_path_violates_writ
         output_spool: None,
         output_spool_max_bytes: csa_process::DEFAULT_SPOOL_MAX_BYTES,
         output_spool_keep_rotated: csa_process::DEFAULT_SPOOL_KEEP_ROTATED,
+        error_marker_scan_enabled: true,
         setting_sources: None,
         sandbox: Some(&sandbox),
         thinking_budget: None,
@@ -516,6 +520,7 @@ async fn test_execute_fails_fast_when_symlinked_shared_npm_cache_resolves_outsid
         output_spool: None,
         output_spool_max_bytes: csa_process::DEFAULT_SPOOL_MAX_BYTES,
         output_spool_keep_rotated: csa_process::DEFAULT_SPOOL_KEEP_ROTATED,
+        error_marker_scan_enabled: true,
         setting_sources: None,
         sandbox: Some(&sandbox),
         thinking_budget: None,
@@ -650,6 +655,7 @@ async fn test_legacy_execute_fails_fast_when_symlinked_shared_npm_cache_resolves
         output_spool: None,
         output_spool_max_bytes: csa_process::DEFAULT_SPOOL_MAX_BYTES,
         output_spool_keep_rotated: csa_process::DEFAULT_SPOOL_KEEP_ROTATED,
+        error_marker_scan_enabled: true,
         setting_sources: None,
         sandbox: Some(&sandbox),
         thinking_budget: None,
