+++ crates/csa-executor/src/transport_tests_tail.rs
@@ -727,6 +727,7 @@ async fn test_execute_stops_after_max_attempts_and_returns_last_failure() {
         output_spool: None,
         output_spool_max_bytes: csa_process::DEFAULT_SPOOL_MAX_BYTES,
         output_spool_keep_rotated: csa_process::DEFAULT_SPOOL_KEEP_ROTATED,
+        error_marker_scan_enabled: true,
         setting_sources: None,
         sandbox: None,
         thinking_budget: None,
